refactor(1013): extract DatabaseSchemaUtils to src/db (SC-001 position 38)
Cat B fresh extraction of DatabaseSchemaUtils (34 refs, 179 LOC) from
MistHelper.py to src/db/database_schema_utils.py using the SC-001 facade
pattern with lazy MistHelper import inside helpers that read live globals.

- Static-method DDL builder class relocated verbatim; single live-global
  reference (ENDPOINT_PRIMARY_KEY_STRATEGIES) resolved via
  importlib.import_module("MistHelper") inside get_endpoint_strategy and
  _build_default_strategy.
- Re-export alias in MistHelper.py preserves MistHelper.DatabaseSchemaUtils
  attribute access for existing callers.
- Coverage omit entry added for the new module; mypy Wave 2 override
  already covers src.db.* (pyproject.toml:162).
- Local gate: black + ruff + py_compile clean; --test baseline 66/0/131.
