+++ pyproject.toml
@@ -238,6 +238,7 @@ omit = [
     "src/maps/*",
     "src/analytics/data_collection_manager.py",
     "src/api/api_data_fetcher.py",
+    "src/db/database_schema_utils.py",
     "src/export/const_definitions_exporter.py",
     "src/export/gateway_test_exporter.py",
     "src/export/license_export_utils.py",
